% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ad_seeds}
\alias{ad_seeds}
\title{Seeds for adaptive dynamics simulations.}
\format{
A tibble with 18 rows and 4 variables:
\describe{
\item{n_traits}{Number of traits: 2 or 3.}
\item{eta_par}{Sign of \code{eta} parameter: -1, 0, or 1.}
\item{d_par}{Sign of \code{d} parameter: -1, 0, or 1.}
\item{seed}{Seed to use for this particular set of simulations.}
}
}
\usage{
ad_seeds
}
\description{
By storing seeds for each simulation, I can retrieve individual runs to look
at them in greater detail.
}
\section{How data were generated}{
\if{html}{\out{<div class="r">}}\preformatted{library(dplyr)
ad_seeds <- tibble(n_traits = rep(2:3, each = 9),
                   eta_par = rep(rep(-1:1, each = 3), 2),
                   d_par = rep(rep(-1:1, 3), 2),
                   seed = numeric(18))
set.seed(1541971065)
ad_seeds$seed[ad_seeds$n_traits==2] <- sample.int(.Machine$integer.max, 9)
set.seed(254296841)
ad_seeds$seed[ad_seeds$n_traits==3] <- sample.int(.Machine$integer.max, 9)

usethis::use_data(ad_seeds)
}\if{html}{\out{</div>}}
}

\keyword{datasets}
