% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dyn.R
\name{adapt_dyn}
\alias{adapt_dyn}
\title{Adaptive dynamics.}
\usage{
adapt_dyn(
  eta,
  d,
  q,
  n_reps,
  n = 1,
  V0 = 1,
  N0 = rep(1, n),
  f = 0.1,
  a0 = 1e-04,
  r0 = 0.5,
  sigma_V0 = 1,
  sigma_N = 0,
  sigma_V = 0,
  max_t = 10000L,
  min_N = 1,
  save_every = 10L,
  mut_sd = 0.1,
  mut_prob = 0.01,
  max_clones = 10000,
  show_progress = TRUE,
  n_threads = 1
)
}
\arguments{
\item{eta}{Number(s) representing the non-additive effects of traits on the
growth rate.
Should be a single number or a symmetrical, numeric, \code{q} by \code{q} matrix.}

\item{d}{Number(s) that adjusts how the focal line is affected by
other lines' trait values.
Should be of length 1 or the same as the number of traits.
If \code{d < 0}, then increases in \code{V_j} (trait that reduces competition
experienced by clone \code{j}) increases competition experienced by clone \code{i},
thereby giving conflicting coevolution.
Conversely, if \code{d > 0}, then increases in \code{V_j} decrease competition
experienced by clone \code{i}, leading to nonconflicting coevolution.}

\item{V0}{Trait value(s) for each starting clone.
For only one starting line, must be a numeric vector or a single
matrix row or column.}

\item{N0}{Abundance(s) for each starting clone. Must be a numeric vector or a single
matrix row or column.}

\item{f}{A single number representing the cost of the trait on the growth rate.}

\item{a0}{A single number representing the base density dependence.}

\item{r0}{A single number representing the base growth rate.}

\item{max_t}{Maximum time simulated.}

\item{min_N}{Minimum N that's considered extant.}

\item{save_every}{Number of time steps between when saving information for output.}

\item{mut_sd}{Standard deviation of the normal distribution used to generate
mutated (i.e., daughter) trait values.}

\item{mut_prob}{Probability of a mutation (creating a daughter lineage).}

\item{max_clones}{Maximum number of clones predicted. This is used only to reserve
memory for some of the inner C++ objects, so when deciding on a value for this,
you should choose a high value.}

\item{show_progress}{Boolean for whether to show a progress bar.}
}
\description{
Adaptive dynamics.
}
