% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dyn.R
\name{adapt_dyn_args}
\alias{adapt_dyn_args}
\title{Create list of arguments for running a set of simulations using \code{adapt_dyn}.}
\usage{
adapt_dyn_args(eta_sign, d_sign, q, ...)
}
\arguments{
\item{eta_sign}{Sign of the \code{eta} parameter.}

\item{d_sign}{Sign of the \code{d} parameter.}

\item{q}{Number of traits.}

\item{...}{Other arguments to add for \code{adapt_dyn} function, namely \code{n_reps}.}
}
\description{
Create list of arguments for running a set of simulations using \code{adapt_dyn}.
}
