% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{add_perterbation}
\alias{add_perterbation}
\title{Add a perturbation to an \code{adaptive_dynamics} object.}
\usage{
add_perterbation(ad_obj, new_prop, new_trait_means, new_N_mean, new_N_sd,
  which_traits = NULL, new_trait_sigmas = NULL, ...)
}
\arguments{
\item{ad_obj}{\code{adaptive_dynamics} object.}

\item{new_prop}{Proportion of new clones.}

\item{new_trait_means}{Mean(s) of new clones' trait(s).}

\item{new_N_mean}{Mean of new clones' abundances.}

\item{new_N_sd}{SD of new clones' abundances.}

\item{which_traits}{Traits to change. Defaults to \code{1:length(new_trait_means)}.}

\item{new_trait_sigmas}{SDs of new clones' trait(s).
Defaults to SD of each trait at the end of the previous run of
\code{adaptive_dynamics()}.}

\item{...}{Arguments to pass to \code{adaptive_dynamics()}.}
}
\description{
Add a perturbation to an \code{adaptive_dynamics} object.
}
