% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{F_t}
\alias{F_t}
\title{Fitness at time t.}
\usage{
F_t(V, N, f, g, C, r0, d)
}
\arguments{
\item{V}{List of row vectors, each containing trait values at time t (1x2 vector)
for a particular clone.}

\item{N}{Row vector of population abundances at time t.}

\item{f}{Effect of traits on growth rate.}

\item{g}{Effect of traits on density dependence.}

\item{C}{Matrix containing non-additive effects of traits on growth rate.}

\item{r0}{Starting growth rate.}

\item{d}{Changes how the focal line is affected by other lines' trait values.
If \code{d < 0}, then increases in \code{V_j} (trait that reduces competition
experienced by clone \code{j}) increases competition experienced by clone \code{i},
thereby giving conflicting coevolution.
Conversely, if \code{d > 0}, then increases in \code{V_j} decrease competition
experienced by clone \code{i}, leading to nonconflicting coevolution.}
}
\description{
Fitness at time t.
}
