% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_gen.R
\name{jacobians}
\alias{jacobians}
\title{Jacobian matrices, one per rep.}
\usage{
jacobians(qg_obj, evo_only = FALSE)
}
\arguments{
\item{qg_obj}{A \code{quant_gen} object from \code{quant_gen} function.}
}
\description{
Jacobian matrices, one per rep.
}
