% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perturb}
\alias{perturb}
\title{Generic method to perturb an object.}
\usage{
perturb(obj, ...)
}
\arguments{
\item{obj}{The object to perturb.}

\item{...}{Additional arguments.}
}
\description{
Implemented only for \code{adapt_dyn} objects thus far.
}
