% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_gen.R
\name{pop_sizes}
\alias{pop_sizes}
\title{Population abundances from quantitative genetics analytical solutions.}
\usage{
pop_sizes(n, eta, d, ...)
}
\arguments{
\item{eta}{Number(s) representing the non-additive effects of traits on the
growth rate.
Should be a single number or a symmetrical, numeric, \code{q} by \code{q} matrix.}

\item{d}{Number(s) that adjusts how the focal line is affected by
other lines' trait values.
Should be of length 1 or the same as the number of traits.
If \code{d < 0}, then increases in \code{V_j} (trait that reduces competition
experienced by clone \code{j}) increases competition experienced by clone \code{i},
thereby giving conflicting coevolution.
Conversely, if \code{d > 0}, then increases in \code{V_j} decrease competition
experienced by clone \code{i}, leading to nonconflicting coevolution.}

\item{...}{Arguments to pass to \code{adapt_dyn()}.}
}
\value{
A numeric vector of abundances, of length \code{n}.
}
\description{
\emph{Note:} Only works for 2 traits for now.
}
