% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_gen.R
\name{quant_gen}
\alias{quant_gen}
\alias{perturb.quant_gen}
\title{Quantitative genetics.}
\usage{
quant_gen(
  eta,
  d,
  q,
  n = 10,
  V0 = NULL,
  N0 = rep(1, n),
  f = 0.1,
  a0 = 0.5,
  r0 = 0.5,
  add_var = rep(0.01, n),
  sigma_V0 = 1,
  sigma_N = 0,
  sigma_V = 0,
  n_reps = 10,
  spp_gap_t = 5000L,
  final_t = 20000L,
  min_N = 1e-04,
  adjust_mu_V = FALSE,
  lnorm_V = TRUE,
  save_every = 10L,
  show_progress = TRUE,
  n_threads = 1
)

\method{perturb}{quant_gen}(
  obj,
  final_t,
  save_every,
  spp_gap_t = 0,
  sigma_V0 = 0,
  new_V = NULL,
  new_Vp = NULL,
  new_N = NULL,
  which_rep = 1,
  ...
)
}
\arguments{
\item{eta}{Number(s) representing the non-additive effects of traits on the
growth rate.
Should be a single number or a symmetrical, numeric, \code{q} by \code{q} matrix.}

\item{d}{Number(s) that adjusts how the focal line is affected by
other lines' trait values.
Should be of length 1 or the same as the number of traits.
If \code{d < 0}, then increases in \code{V_j} (trait that reduces competition
experienced by clone \code{j}) increases competition experienced by clone \code{i},
thereby giving conflicting coevolution.
Conversely, if \code{d > 0}, then increases in \code{V_j} decrease competition
experienced by clone \code{i}, leading to nonconflicting coevolution.}

\item{V0}{Trait value(s) for each starting clone.
For only one starting line, must be a numeric vector or a single
matrix row or column.}

\item{N0}{Abundance(s) for each starting clone. Must be a numeric vector or a single
matrix row or column.}

\item{f}{A single number representing the cost of the trait on the growth rate.}

\item{a0}{A single number representing the base density dependence.}

\item{r0}{A single number representing the base growth rate.}

\item{add_var}{Vector of additive genetic variances for all starting species.}

\item{sigma_V0}{Standard deviation for normal distribution from which
starting trait values can be derived.
Set to 0 for species to start with the exact values of traits
specified in \code{V0}.}

\item{sigma_N}{Standard deviation for stochasticity in population dynamics.}

\item{sigma_V}{Standard deviation for stochasticity in trait evolution.}

\item{n_reps}{Number of reps to perform.}

\item{spp_gap_t}{Time period between each species introduction.}

\item{final_t}{Length of final time period where all species are together.}

\item{min_N}{Minimum N that's considered extant.}

\item{adjust_mu_V}{Logical for whether to adjust the mean of the lognormal
distribution that creates stochasticity in trait values so that
the mean of the transformed distribution (i.e.,
\code{exp(~ N(mu, sigma_V))}) is 1.}

\item{lnorm_V}{Logical for whether the distribution for stochasticity in
evolution should be lognormal. The alternative is a normal distribution.
Defaults to \code{TRUE}.}

\item{save_every}{Number of time steps between when saving information for output.}

\item{show_progress}{Boolean for whether to show a progress bar.}

\item{n_threads}{Number of cores to use. Defaults to 1.}

\item{obj}{A \code{quant_gen} object.}

\item{new_V}{List of trait values for the new species being added.
Defaults to \code{NULL}, which causes no new species to be add.}

\item{new_N}{Numeric vector of new species to add. Defaults to \code{NULL}, which
causes no new species to be add.}

\item{which_rep}{Integer for which repetition to perturb.}

\item{...}{Other parameters to pass to \code{quant_gen}.}
}
\value{
A \code{quant_gen} object with \code{nv} (for N and V output) and
\code{call} (for original call) fields.
}
\description{
Quantitative genetics.
}
\section{Methods (by generic)}{
\itemize{
\item \code{perturb}: Add a perturbation to a \code{quant_gen} object.
}}

