% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_gen.R
\name{quant_gen_args}
\alias{quant_gen_args}
\title{Create list of arguments for running a set of simulations using \code{quant_gen}.}
\usage{
quant_gen_args(eta_sign, d_sign, q)
}
\arguments{
\item{eta_sign}{Sign of the \code{eta} parameter.}

\item{d_sign}{Sign of the \code{d} parameter.}

\item{q}{Number of traits.}
}
\description{
Create list of arguments for running a set of simulations using \code{quant_gen}.
}
