% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_gen.R
\name{stable_points}
\alias{stable_points}
\alias{unstable_points}
\title{Stable points from quantitative genetics simulations and analytical solutions.}
\usage{
stable_points(
  eta,
  f = 0.1,
  a0 = 0.5,
  r0 = 0.5,
  return_geom = FALSE,
  line_n = 1000,
  ...
)

unstable_points(eta, f = 0.1, a0 = 0.5, r0 = 0.5, return_geom = FALSE, ...)
}
\arguments{
\item{eta}{Number(s) representing the non-additive effects of traits on the
growth rate.
Should be a single number or a symmetrical, numeric, \code{q} by \code{q} matrix.}

\item{f}{A single number representing the cost of the trait on the growth rate.}

\item{a0}{A single number representing the base density dependence.}

\item{r0}{A single number representing the base growth rate.}

\item{return_geom}{Logical for whether to return a \code{ggplot2} \verb{geom_*} object for
plotting rather than a \code{tibble}. Defaults to \code{FALSE}.}

\item{line_n}{Number of points to use for lines. Defaults to \code{1000}.}

\item{...}{Other arguments to pass to the call to \code{geom_point} or \code{geom_path}.}
}
\value{
A \code{tibble} if \code{return_geom} is \code{FALSE}, and a  \verb{geom_*} object if it's \code{TRUE}.
}
\description{
\emph{Note:} Only works for 2 traits for now.
}
\section{Functions}{
\itemize{
\item \code{unstable_points}: Unstable points
}}

