% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_black}
\alias{theme_black}
\title{Black theme to use for presentations}
\usage{
theme_black(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Implemented only for \code{adapt_dyn} objects thus far.
}
