% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trnorm}
\alias{trnorm}
\title{Random deviates from truncated normal distribution}
\usage{
trnorm(n, mu, sigma)
}
\arguments{
\item{n}{Number of deviates.}

\item{mu}{Mean of distribution.}

\item{sigma}{Standard deviation of distribution.}
}
\value{

}
\description{
Note: truncated at zero.
}
