% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_gen.R
\name{quant_gen}
\alias{quant_gen}
\title{Quantitative genetics.}
\usage{
quant_gen(
  eta,
  d,
  q,
  n = 10,
  V0 = 1,
  N0 = rep(1, n),
  f = 0.1,
  a0 = 1e-04,
  r0 = 0.5,
  add_var = rep(0.01, n),
  sigma_V0 = 1,
  sigma_N = 0,
  sigma_V = 0,
  phenos = TRUE,
  n_reps = 10,
  spp_gap_t = 500L,
  final_t = 5000L,
  min_N = 1,
  save_every = 10L,
  show_progress = TRUE,
  n_threads = 1
)
}
\arguments{
\item{eta}{Number(s) representing the non-additive effects of traits on the
growth rate.
Should be a single number or a symmetrical, numeric, \code{q} by \code{q} matrix.}

\item{d}{Number(s) that adjusts how the focal line is affected by
other lines' trait values.
Should be of length 1 or the same as the number of traits.
If \code{d < 0}, then increases in \code{V_j} (trait that reduces competition
experienced by clone \code{j}) increases competition experienced by clone \code{i},
thereby giving conflicting coevolution.
Conversely, if \code{d > 0}, then increases in \code{V_j} decrease competition
experienced by clone \code{i}, leading to nonconflicting coevolution.}

\item{V0}{Trait value(s) for each starting clone.
For only one starting line, must be a numeric vector or a single
matrix row or column.}

\item{N0}{Abundance(s) for each starting clone. Must be a numeric vector or a single
matrix row or column.}

\item{f}{A single number representing the cost of the trait on the growth rate.}

\item{a0}{A single number representing the base density dependence.}

\item{r0}{A single number representing the base growth rate.}

\item{add_var}{Vector of additive genetic variances for all starting species.}

\item{sigma_V0}{Standard deviation for normal distribution from which
starting axis values can be derived.
Set to 0 for species to start with the exact values of axes
specified in \code{V0}.}

\item{sigma_N}{Standard deviation for stochasticity in population dynamics.}

\item{sigma_V}{Standard deviation for stochasticity in axis evolution.}

\item{phenos}{Logical for whether axis evolution stochasticity produces
phenotypes or whether it just adds variance to the genotype directly.
Defaults to \code{TRUE}.}

\item{n_reps}{Number of reps to perform.}

\item{spp_gap_t}{Time period between each species introduction.}

\item{final_t}{Length of final time period where all species are together.}

\item{min_N}{Minimum N that's considered extant.}

\item{save_every}{Number of time steps between when saving information for output.}

\item{show_progress}{Boolean for whether to show a progress bar.}

\item{n_threads}{Number of cores to use. Defaults to 1.}
}
\value{
A \code{quant_gen} object with \code{nv} (for N and V output) and
\code{call} (for original call) fields.
}
\description{
Quantitative genetics.
}
