% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_gen.R
\name{quant_gen}
\alias{quant_gen}
\title{Quantitative genetics.}
\usage{
quant_gen(
  eta,
  d,
  q,
  n = 100,
  V0 = rep(list(matrix(0, 1, q)), n),
  N0 = rep(1, n),
  f = 0.1,
  a0 = 0.5,
  r0 = 0.5,
  add_var = rep(0.5, n),
  perturb_sd = 1,
  n_reps = 100,
  start_t = 0,
  max_t = 1000000L,
  min_N = 1e-04,
  save_every = 10000L,
  show_progress = TRUE,
  n_threads = 1
)
}
\arguments{
\item{eta}{Number(s) representing the non-additive effects of traits on the
growth rate.
Should be a single number or a symmetrical, numeric, \code{q} by \code{q} matrix.}

\item{d}{Number(s) that adjusts how the focal line is affected by
other lines' trait values.
Should be of length 1 or the same as the number of traits.
If \code{d < 0}, then increases in \code{V_j} (trait that reduces competition
experienced by clone \code{j}) increases competition experienced by clone \code{i},
thereby giving conflicting coevolution.
Conversely, if \code{d > 0}, then increases in \code{V_j} decrease competition
experienced by clone \code{i}, leading to nonconflicting coevolution.}

\item{V0}{Trait value(s) for each starting clone.
For only one starting line, must be a numeric vector or a single
matrix row or column.}

\item{N0}{Abundance(s) for each starting clone. Must be a numeric vector or a single
matrix row or column.}

\item{f}{A single number representing the cost of the trait on the growth rate.}

\item{a0}{A single number representing the base density dependence.}

\item{r0}{A single number representing the base growth rate.}

\item{add_var}{Vector of additive genetic variances for all starting species.}

\item{perturb_sd}{Standard deviation of the perturbation.}

\item{n_reps}{Number of reps to perform.}

\item{start_t}{Number of starting iterations (before the perturbation).}

\item{max_t}{Maximum time simulated.}

\item{min_N}{Minimum N that's considered extant.}

\item{save_every}{Number of time steps between when saving information for output.}

\item{show_progress}{Boolean for whether to show a progress bar.}

\item{n_threads}{Number of cores to use. Defaults to 1.}
}
\value{
A \code{quant_gen} object with \code{nv} (for N and V output) and \code{fs} (for fitness
and selection output) fields.
}
\description{
Quantitative genetics.
}
